move: 페이지 스타일 파일 위치 이동
- components/styles/pages 하위로 이동
