+++ components/styles/pages/main.style.ts
