+++ components/styles/pages/reward.style.ts
