+++ components/styles/pages/welcome.style.ts
