+++ pages/index.tsx
@@ -7,8 +7,7 @@ import Header from '@/components/Header';
 import Container from '@/components/Container';
 import ReportList from '@/components/ReportList';
 import { ROUTES } from '@/constants';
-import * as styles from '@/components/styles/main.style';
-
+import * as styles from '@/components/styles/pages/main.style';
 
 export default function Home() {
   const { data: reports } = useQuery({
