+++ pages/welcome/index.tsx
@@ -4,7 +4,7 @@ import HeadMeta from '@/components/HeadMeta';
 import Container from '@/components/Container';
 import Button from '@/components/Button';
 import { ROUTES } from '@/constants';
-import * as styles from '@/components/styles/welcome.style';
+import * as styles from '@/components/styles/pages/welcome.style';
 
 export default function Welcome() {
   const router = useRouter();
